LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/pixel-manipulation/2d.imageData.put.clip-expected.txt

2d.imageData.put.clip

putImageData() is not affected by clipping regions


PASS putImageData() is not affected by clipping regions 
